feat: code clean up!
