+++ package.json
@@ -1,5 +1,5 @@
 {
-	"name": "Project Mage",
+	"name": "ProjectMage",
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
